# services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Seckill Mall Commerce API
  version: 0.4.0
  description: 阶段 4 商城微服务 API。Gateway 显式路由 Identity、Catalog、Cart、Order、Payment 与 Fulfillment；金额字段均使用人民币分。
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: API Gateway
...
paths:
...
  /orders:
    get:
      tags: [Order]
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 当前用户订单分页
    post:
      tags: [Order]
      operationId: createOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id]
              properties:
                address_id:
                  type: integer
                  format: int64
                  minimum: 1
                items:
                  type: array
                  description: 可选；为空时 Gateway 读取当前用户 Cart
                  items:
                    $ref: '#/components/schemas/CartItemInput'
      responses:
        '201':
          description: 已创建待支付订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '200':
          description: 返回同一幂等订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '409':
          $ref: '#/components/responses/Error'
  /seckill/orders:
    post:
      tags: [Order]
      operationId: createSeckillOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address_id, sku_id, quantity]
              properties:
                address_id:
                  type: integer
                  format: int64
                sku_id:
                  type: integer
                  format: int64
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 99
      responses:
        '201':
          description: 秒杀订单进入统一待支付状态
        '409':
          $ref: '#/components/responses/Error'
  /orders/{order_id}:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    get:
      tags: [Order]
      operationId: getOrder
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 订单详情、状态历史、支付与物流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderEnvelope'
        '404':
          $ref: '#/components/responses/Error'
  /orders/{order_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Order]
      operationId: cancelOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 255
      responses:
        '200':
          description: 已取消并释放库存
        '409':
          $ref: '#/components/responses/Error'
...
  /orders/{order_id}/refunds:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Order]
      operationId: refundOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  maxLength: 255
      responses:
        '200':
          description: Mock 退款已完成
        '409':
          $ref: '#/components/responses/Error'
  /orders/{order_id}/payments:
    parameters:
      - $ref: '#/components/parameters/OrderID'
    post:
      tags: [Payment]
      operationId: createMockPayment
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Mock 支付单和仅用于测试的回调签名
        '200':
          description: 返回同一订单的幂等支付单
  /payments/callback:
    post:
      tags: [Payment]
      operationId: completeMockPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_no, callback_ref, signature]
              properties:
                payment_no:
                  type: string
                callback_ref:
                  type: string
                signature:
                  type: string
      responses:
        '200':
          description: 幂等完成支付并确认库存
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderID:
      name: order_id
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
...
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
        request_id:
          type: string
...
    CartItemInput:
      type: object
      required: [sku_id, quantity]
      properties:
        sku_id:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 99
...
    Order:
      type: object
      required: [order_id, user_id, order_type, status, total_amount_cents]
      properties:
        order_id:
          type: string
        user_id:
          type: integer
          format: int64
        order_type:
          type: string
          enum: [normal, seckill]
        status:
          type: string
          enum: [pending_payment, paid, shipped, completed, canceled, refund_pending, refunded]
        total_amount_cents:
          type: integer
          format: int64
          minimum: 0
    OrderEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Order'
...